+++ grails-app/services/io/xh/hoist/alertbanner/AlertBannerService.groovy
@@ -33,6 +33,8 @@ class AlertBannerService extends BaseService {
     private final static String blobType = 'xhAlertBanner';
     private final static String blobOwner = 'xhAlertBannerService';
 
+    private final static String presetsBlobName = 'xhAlertBannerPresets';
+
     private final emptyAlert = [active: false]
     private Map cachedBanner = emptyAlert
 
@@ -69,6 +71,24 @@ class AlertBannerService extends BaseService {
         }
         refreshCachedBanner()
     }
+
+    List getAlertPresets() {
+        def svc = jsonBlobService,
+            presetsBlob = svc.list(blobType, blobOwner).find { it.name == presetsBlobName }
+
+        presetsBlob ? JSONParser.parseArray(presetsBlob.value) : []
+    }
+
+    void setAlertPresets(List presets) {
+        def svc = jsonBlobService,
+            blob = svc.list(blobType, blobOwner).find { it.name == presetsBlobName }
+        if (blob) {
+            svc.update(blob.token, [value: presets], blobOwner)
+        } else {
+            svc.create([type: blobType, name: presetsBlobName, value: presets], blobOwner)
+        }
+    }
+
 
     //----------------------------
     // Implementation